Fixed linting

+++ core/services/ocrcommon/transmitter.go
@@ -75,7 +75,7 @@ func (t *transmitter) CreateEthTransaction(ctx context.Context, toAddress common
 		FromAddress:      roundRobinFromAddress,
 		ToAddress:        toAddress,
 		EncodedPayload:   payload,
-		FeeLimit:         uint64(t.gasLimit),
+		FeeLimit:         t.gasLimit,
 		ForwarderAddress: t.forwarderAddress(),
 		Strategy:         t.strategy,
 		Checker:          t.checker,
